+++ sources/scripts/interaction/bookmarks_interface.js
@@ -9,7 +9,8 @@
      "needs_setup", "authenticate", "load",
      "clear", "setup", "change_authentication",
      "lock", "unlock", "is_locked", "is_unlocked",
-     "export_encrypted_data", "export_plain_data"]
+     "export_encrypted_data", "export_plain_data",
+     "import_encrypted_data", "import_plain_data"]
     .forEach(method_name =>
     {
         this_module[method_name] = (...arguments) =>